Split `index.ts` into `patch-apk.ts` and `patch-app-bundle.ts`
